Allow branch labels in node-data JSONs
Previously branch labels could not be specified in data passed to
`augur export v2` except for two "special cases":
(1) Mutations (stored in node-data-json -> nodes) would create branch
labels "aa", if applicable.
(2) `clade_annotation` (stored in node-data-json -> nodes) was
interpreted to be the "clade" branch label, and exported as such.

Here we extend the allowed node-data structure to include a top-level
key `branches` as described in [1] and the test data added here [2].
This data is exported in the appropriate format for Auspice (unchanged).
This provides an way for analyses to export custom branch labels other
than the two special-cases described above. Note that currently no augur
commands can produce such data, but this will change - see [1] for more.

This work also induced two smaller changes. The auspice config JSON
schema is extended the default branch label displayed to be any value.
Secondly, the requirement for node-data JSONs to specify "nodes" has
been relaxed (see [2] for an example); if neither "nodes" nor "branches"
are defined then we raise a validation error.

[1] https://github.com/nextstrain/augur/issues/720
[2] ./tests/functional/export_v2/branch-labels.json