augur clades allows attribute name to be specified
Previously the `augur clades` command produced a node-data JSON
which stored clade membership as the node-attr "clade_membership"
and defined the basal nodes of each clade with the node-attr
"clade_annotation". `augur export v2` interpreted the latter as a
special-case and turned it into a branch label of the same name.

The previous commit allowed `augur export` to be supplied node-data
JSONs with a `branches` dictionary. Here we update `augur clades`
to export data in this structure, which allows the user to specify
the keys to use via the `--attribute-name` arg.

This commit breaks backwards compatibility for pipelines as the default
attribute name is "clade". This will result in dataset (auspice) JSONs
with the same branch labelling as before, but with a different node-attr
(was "clade_membership", now "clade"). As `augur export v2` will make
colorings for all node-attrs in in node-data JSONs, this will be
exported as a "clade" coloring with no changes needed, however auspice
config JSONs may now refer to a non-existent "clade_membership" key.

`augur export v2` has been updated to no longer special-case
`clade_membership` or `clade_annotation` node attrs. We print a
warning if an auspice config JSON refers to `clade_membership` to
help users update their configs.

Functional tests for `augur clades` have been added.

Closes https://github.com/nextstrain/augur/issues/720