[augur ancestral] create annotation block
For a detailed write-up of the bug which motivated this commit, see
https://github.com/nextstrain/augur/issues/881.

By storing the (nucleotide) genome annotation in the node-data produced
from augur-ancestral we make this information available for export.
Previously this information was only exported by `augur translate` which
was problematic for workflows which didn't perform translation.

No changes are needed to `augur export v2` (which may now process
multiple "annotations" blocks) due to the behavior of
`NodeData.deep_add_or_update` which will recurse into dicts in
annotation blocks and when confronted with non-dict values which already
exist overwrite them. This poses a potential problem where two node-data
JSONs which (e.g.) define different `annotations['nuc']` coordinates
will not raise any error and the output coodinates are dependent on
the order the node-data JSONs were provided to `augur export v2`.

Closes https://github.com/nextstrain/augur/issues/881.